Renamed container builder
