+++ kubernetes/container/containers.go
@@ -1,4 +1,4 @@
-package containers
+package container
 
 import (
 	corev1 "k8s.io/api/core/v1"
